+++ typescript/infra/config/environments/mainnet3/warp/warpIds.ts
@@ -148,4 +148,7 @@ export enum WarpRouteIds {
 
   // ctUSD
   CitreaUSD = 'ctUSD/citrea',
+
+  // RISE routes
+  BscEthereumRISE = 'RISE/bsc-ethereum',
 }