+++ typescript/github-proxy/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @hyperlane-xyz/github-proxy
 
+## 30.0.0
+
 ## 29.1.0
 
 ## 29.0.1